+++ include/RequestHandler.hpp
@@ -11,7 +11,7 @@ const std::string get_mime_type(const std::filesystem::path& extension);
 const std::string create_directory_listing(const std::filesystem::path& directory);
 const LocationContext* find_location(const std::string& folder_path, const ServerContext& config);
 Response build_error_page(HTTPStatusCode status_code, const LocationContext* location, const ServerContext& config);
-bool is_cgi_request(const std::string& uri);
+bool is_cgi_request(const std::filesystem::path& root, const std::string& uri);
 bool request_method_allowed(const LocationContext* location, HTTPMethod method);
 std::filesystem::path map_uri(std::string uri, const LocationContext* location, const Root& root);
